crud_operation/requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.2
django-cors-headers==4.6.0
django-dotenv==1.4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
mysqlclient==2.2.6
pillow==11.0.0
platformdirs==4.3.6
PyJWT==2.10.1
sqlparse==0.5.1
tzdata==2024.2
virtualenv==20.27.0
